backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
pydantic-settings
mysql-connector-python
python-multipart